refactor(me): split me.h into api/inline/layout headers
Decompose me.h (~700 LOC) into three focused headers:
- me.h: public API + DRAM types only
- me_inline.h: inline helpers needing instance struct visibility
- me_layout.h: on-disk CXL layout (header/CB/membership/slot)

Move cold-path helpers (me_leave_successor, me_membership_tick_heartbeat)
out of inline header into me.c. Consolidate per-shard arrays into
struct marufs_me_shard and DRAM is_holder fast path. Wire callers
(bootstrap, me_order, me_request, nrht, sysfs) to new layout.
